Update irr_nonauth_cleanup.py
I found some origin's within the radb.db.gz file are origin:asXXXX with no spacing, so I adjusted the code to look for the digits in the origin line via a regex grouping match. These changes allowed me to run this on the python3.8 with the radb.db.gz file.